requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-simpleui==2022.3.19
djangorestframework==3.13.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
psycopg2==2.9.3
pytz==2022.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
tzdata==2022.1
